traffic-classifier/requirements.txt
pandas
numpy
scipy
scikit-learn
pymilvus
openai
watchdog
python-dotenv
tqdm
